+++ lib/braintree/transaction.rb
@@ -92,6 +92,8 @@ module Braintree
 
     attr_reader :add_ons
     attr_reader :additional_processor_response          # The raw response from the processor.
+    # NEXT_MAJOR_VERSION Remove this class.
+    # DEPRECATED The American Express Checkout payment method is deprecated.
     attr_reader :amex_express_checkout_details
     attr_reader :amount
     attr_reader :android_pay_details
@@ -126,6 +128,8 @@ module Braintree
     # DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
     attr_reader :ideal_payment_details
     attr_reader :local_payment_details
+    # NEXT_MAJOR_VERSION Remove this class.
+    # DEPRECATED The Masterpass Card payment method is deprecated.
     attr_reader :masterpass_card_details
     attr_reader :merchant_account_id
     attr_reader :network_response_code                  # Response code from the card network